Move DeleteDialog of CrudContextMenu into a separate file
